+++ src/main/java/com/donkeys_today/server/presentation/auth/dto/response/UserSignUpResponse.java
@@ -1,0 +1,12 @@
+package com.donkeys_today.server.presentation.auth.dto.response;
+
+public record UserSignUpResponse(
+        Long userId,
+        String accessToken,
+        String refreshToken
+) {
+
+    public static UserSignUpResponse of(Long userId, String accessToken, String refreshToken) {
+        return new UserSignUpResponse(userId, accessToken, refreshToken);
+    }
+}
