--- src/main/java/com/donkeys_today/server/presentation/user/dto/requset/UserSignInRequest.java
@@ -1,12 +1,0 @@
-package com.donkeys_today.server.presentation.user.dto.requset;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-public record UserSignInRequest(
-    @Schema(description = "플랫폼", example = "apple/kakao")
-    String platform,
-    @Schema(description = "애플 로그인", example = "ey123fg.asdafa.asddsa")
-    String id_token
-) {
-
-}
