--- src/main/java/com/donkeys_today/server/presentation/user/dto/response/UserSignInResponse.java
@@ -1,11 +1,0 @@
-package com.donkeys_today.server.presentation.user.dto.response;
-
-public record UserSignInResponse(
-    Long userId,
-    String accessToken,
-    String refreshToken
-) {
-  public static UserSignInResponse of(Long userId, String accessToken, String refreshToken){
-    return new UserSignInResponse(userId, accessToken, refreshToken);
-  }
-}
