--- src/main/java/com/donkeys_today/server/presentation/user/dto/response/UserSignUpResponse.java
@@ -1,12 +1,0 @@
-package com.donkeys_today.server.presentation.user.dto.response;
-
-public record UserSignUpResponse(
-    Long userId,
-    String accessToken,
-    String refreshToken
-) {
-
-  public static UserSignUpResponse of(Long userId, String accessToken, String refreshToken){
-    return new UserSignUpResponse(userId, accessToken, refreshToken);
-  }
-}
